refactor(lint-framework): rename from `harper-content-framework`
