+++ justfile
@@ -24,12 +24,12 @@ build-harperjs: build-wasm
   # Generate API reference
   ./docs.sh
 
-# Build the browser content framework module
-build-content-framework:
+# Build the browser lint framework module
+build-lint-framework:
   #!/usr/bin/env bash
   set -eo pipefail
 
-  cd "{{justfile_directory()}}/packages/harper-content-framework"
+  cd "{{justfile_directory()}}/packages/lint-framework"
   pnpm install
   pnpm build
 
@@ -75,7 +75,7 @@ build-wp: build-harperjs
   pnpm plugin-zip
 
 # Compile the website's dependencies and start a development server. Note that if you make changes to `harper-wasm`, you will have to re-run this command.
-dev-web: build-harperjs build-content-framework
+dev-web: build-harperjs build-lint-framework
   #!/usr/bin/env bash
   set -eo pipefail
 
@@ -84,7 +84,7 @@ dev-web: build-harperjs build-content-framework
   pnpm dev
 
 # Build the Harper website.
-build-web: build-harperjs build-content-framework
+build-web: build-harperjs build-lint-framework
   #!/usr/bin/env bash
   set -eo pipefail
   
@@ -105,7 +105,7 @@ build-obsidian: build-harperjs
   zip harper-obsidian-plugin.zip manifest.json main.js
 
 # Build the Chrome extension.
-build-chrome-plugin: build-harperjs build-content-framework
+build-chrome-plugin: build-harperjs build-lint-framework
   #!/usr/bin/env bash
   set -eo pipefail
   
@@ -115,7 +115,7 @@ build-chrome-plugin: build-harperjs build-content-framework
   pnpm zip-for-chrome
 
 # Start a development server for the Chrome extension.
-dev-chrome-plugin: build-harperjs build-content-framework
+dev-chrome-plugin: build-harperjs build-lint-framework
   #!/usr/bin/env bash
   set -eo pipefail
   
@@ -125,7 +125,7 @@ dev-chrome-plugin: build-harperjs build-content-framework
   pnpm dev
 
 # Build the Firefox extension.
-build-firefox-plugin: build-harperjs build-content-framework
+build-firefox-plugin: build-harperjs build-lint-framework
   #!/usr/bin/env bash
   set -eo pipefail
