+++ packages/chrome-plugin/package.json
@@ -52,7 +52,7 @@
 		"@tailwindcss/vite": "^4.1.4",
 		"@webcomponents/custom-elements": "^1.6.0",
 		"harper.js": "workspace:*",
-		"harper-content-framework": "workspace:*",
+		"lint-framework": "workspace:*",
 		"jquery": "^3.7.1",
 		"lodash-es": "^4.17.21",
 		"lru-cache": "^11.1.0",
