--- packages/harper-content-framework/.gitignore
@@ -1,1 +1,0 @@
-dist
