+++ packages/lint-framework/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "harper-content-framework",
+	"name": "lint-framework",
 	"version": "0.0.1",
 	"license": "Apache-2.0",
 	"type": "module",
