+++ packages/lint-framework/src/assets/bookDownSvg.ts
