+++ packages/lint-framework/src/lint/lintKindColor.ts
