+++ packages/lint-framework/vite.config.ts
@@ -6,7 +6,7 @@ export default defineConfig({
 	build: {
 		lib: {
 			entry: resolve(__dirname, 'src/index.ts'),
-			name: 'harperContentFramework',
+			name: 'lintFramework',
 			fileName: 'index',
 			formats: ['es'],
 		},
